% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_compute.R
\name{A_compute}
\alias{A_compute}
\title{A_compute: Compute adjacency matrix to construct a network}
\usage{
A_compute(df, struct = NULL)
}
\arguments{
\item{df}{a data frame containing a column ISRC codes, and a column with the registrant of each ISRC}

\item{struct}{a parameter that allows you to specify the name of the column should be used to create the network structure. If NULL, the structure will be based on the registrant.}
}
\value{
returns an adjacency matrix
}
\description{
A_compute: Compute adjacency matrix to construct a network
}
\examples{
library(openmuse)

A <- A_compute(example)

#Using the struct parameter
A <- A_compute(example, "registrant")


}
