% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{example}
\alias{example}
\title{My Dataset}
\format{
A data frame with 1204 rows and 6 variables:
\describe{
\item{id}{The track Spotify id.}
\item{isrc}{The ISRC of the track}
\item{registrant}{The registrant code extracted from the ISRC}
\item{year}{The year of reference extracted from the ISRC}
\item{designation}{The designation code extracted from the ISRC}
\item{popularity}{Spotify popularity index of the track at the moment the query was made}
}
}
\source{
\url{https://example.com/dataset-source}
}
\usage{
example
}
\description{
This data set is used to showcase the sampling algorithm.
It contains 6 variables and it was collected on June 24, 2024.
}
\keyword{datasets}
